stackdriver: Use package with RPC for MQL query
This uses the Cloud Monitoring package that has support for MQL queries, which
will be useful when we start supporting any percentile for latency. The package
used here is deprecated, so we should migrate to the new package once support
for MQL queries is added to the new Cloud Monitoring package.

Signed-off-by: Getulio Valentin Sánchez <gvso@google.com>