metrics: Remove the Query interface
The Query interface was always passed through the chain without changes to it,
which makes it very hard to deduct its use. This removes the interface
altogether and keep the filters as parameters of the provider constructor.

A new method SetCandidateRevision was added to the Provider interface since the
candidate revision can only be determined during the rollout process.
Furthermore, it looks like knowing the queried resource (i.e. the candidate
revision) is necessary for any metrics provider.

Signed-off-by: Getulio Valentin Sánchez <gvso@google.com>