main: Print all health criteria
This modifies printHealthCriteria to make it print all the health criteria
passed to it. That is, it removes the switch statement with a default case
printing "invalid criterion". This is useful since once more metrics are
supported, we don't need to worry about this function. Validation is done a few
statements later anyways.

Signed-off-by: Getulio Valentin Sánchez <gvso@google.com>