rollout: Indicate if traffic config was changed
This adds a boolean return value to explicitly indicate if the traffic
configuration was changed during the update of the service or the traffic.
Indicating that the traffic changed would be useful when we add annotations to
all services that have stable and candidate revisions. That is, all calls to
UpdateService are gonna update at least the annotation part given that there are
stable and candidate revisions.

Signed-off-by: Getulio Valentin Sánchez <gvso@google.com>